Create Framework Layer and refactor code
