+++ app/src/main/java/com/godamy/myplayer/data/MediaRepository.kt
@@ -3,12 +3,14 @@ package com.godamy.myplayer.data
 import com.godamy.myplayer.App
 import com.godamy.myplayer.R
 import com.godamy.myplayer.domain.MediaItem
+import com.godamy.myplayer.framework.datasource.MediaItemRoomDataSource
+import com.godamy.myplayer.framework.datasource.MediaItemServerDataSource
 
 class MediaRepository(application: App) {
 
     private val regionRepository = RegionRepository(application)
-    private val localDataSource = MediaLocalDataSource(application.db.mediaItemDao())
-    private val remoteDataSource = MediaRemoteDataSource(application.getString(R.string.api_key))
+    private val localDataSource = MediaItemRoomDataSource(application.db.mediaItemDao())
+    private val remoteDataSource = MediaItemServerDataSource(application.getString(R.string.api_key))
 
     val popularMovies = localDataSource.mediaItems
