+++ app/src/main/java/com/godamy/myplayer/data/datasource/MediaItemRemoteDataSource.kt
@@ -1,0 +1,7 @@
+package com.godamy.myplayer.data.datasource
+
+import com.godamy.myplayer.domain.MediaItem
+
+interface MediaItemRemoteDataSource {
+    suspend fun requestPopularMovies(region: String): List<MediaItem>
+}
