+++ app/src/main/java/com/godamy/myplayer/framework/datasource/PlayServiceLocationDataSource.kt
@@ -1,4 +1,4 @@
-package com.godamy.myplayer.data
+package com.godamy.myplayer.framework.datasource
 
 import android.annotation.SuppressLint
 import android.app.Application
